requirements.txt
# ── Core web framework ────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0

# ── Auth ──────────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# ── HTTP / scraping ───────────────────────────────────────────
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.11

# ── AI / LLM ──────────────────────────────────────────────────
google-genai>=0.8.0
google-generativeai>=0.8.0
openai>=1.30.0
json-repair>=0.25.0

# ── ML (trend engine) ─────────────────────────────────────────
sentence-transformers>=2.7.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0

# ── Media ─────────────────────────────────────────────────────
pillow>=10.3.0

# ── Trend sources ─────────────────────────────────────────────
pytrends>=4.9.2
praw>=7.7.1

# ── Utilities ─────────────────────────────────────────────────
tqdm>=4.66.0
rich>=13.7.0
